Move unhandled-URLs handling from config/run to the CLI reporter
The `unhandledUrls` option no longer lives in the config schema, and
`run()` no longer decides what to do when no plugin handles a URL — it
just passes the null result through. The behavior is now an option of
`cliReporter` and exposed as a `--unhandled-urls` CLI flag on the run
command (validated with typanion), defaulting to "error" as before.
