+++ packages/todone/package.json
@@ -61,6 +61,7 @@
     "dedent": "^1.7.2",
     "globby": "^16.2.0",
     "jiti": "^2.7.0",
+    "typanion": "^3.8.0",
     "zod": "^4.4.3"
   },
   "devDependencies": {
