+++ packages/todone/src/commands/run/impl.ts
@@ -3,11 +3,14 @@ import { loadConfigFile } from "#/lib/config/load";
 import { cliReporter } from "#/lib/reporters/cli";
 import { RunCommand } from "./index";
 
-export default async ({ onlyExpired, locale }: RunCommand) => {
+export default async ({ onlyExpired, locale, unhandledUrls }: RunCommand) => {
   const config = await loadConfigFile();
 
   await run({
     ...config,
-    plugins: [...config.plugins, cliReporter({ onlyExpired, locale })],
+    plugins: [
+      ...config.plugins,
+      cliReporter({ onlyExpired, locale, unhandledUrls }),
+    ],
   });
 };
