+++ packages/todone/src/index.ts
@@ -6,20 +6,6 @@ import { getFiles } from "./lib/files";
 import type { Plugin } from "./plugin";
 import type * as t from "./types";
 
-class UnhandledUrlError extends Error {
-  constructor(match: t.Match) {
-    const {
-      url,
-      file,
-      position: { line, column },
-    } = match;
-    super(
-      `No plugin returned a result for ${url} (${file.localPath}:${line}:${column}). ` +
-        `Add a plugin that handles this URL, or set \`unhandledUrls: "warn"\` or \`"ignore"\` in your todone config.`,
-    );
-  }
-}
-
 export interface RunOptions {
   /**
    * If set, all reporting goes to this plugin instead of the configured ones.
@@ -30,7 +16,7 @@ export interface RunOptions {
 }
 
 export const run = async (
-  { globs, gitignore, keyword, plugins, unhandledUrls }: Config,
+  { globs, gitignore, keyword, plugins }: Config,
   { forcedReporter }: RunOptions = {},
 ) => {
   await using container = new PluginContainer(plugins);
@@ -40,27 +26,11 @@ export const run = async (
 
   const reporter = forcedReporterContainer ?? container;
 
-  const check = async (match: t.Match): Promise<t.Result> => {
-    const result = await container.checkMatch({ url: match.url });
-
-    if (result === null) {
-      switch (unhandledUrls) {
-        case "error":
-          throw new UnhandledUrlError(match);
-        case "warn":
-          await reporter.warn(
-            `no plugin handled ${match.url} (${match.file.localPath}:${match.position.line}:${match.position.column})`,
-          );
-        // fallthrough
-        case "ignore":
-          break;
-        default:
-          return unhandledUrls satisfies never;
-      }
-    }
-
-    return { url: match.url, match, result };
-  };
+  const check = async (match: t.Match): Promise<t.Result> => ({
+    url: match.url,
+    match,
+    result: await container.checkMatch({ url: match.url }),
+  });
 
   const results = await it
     .from(getFiles(globs, { cwd: process.cwd(), gitignore: gitignore }))
